backend/foodgram/urls.py
from django.conf.urls.static import static
from django.contrib import admin
from django.urls import path, include
from rest_framework.routers import DefaultRouter

from .settings import MEDIA_URL, MEDIA_ROOT

router = DefaultRouter()

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include('api.urls')),
]

urlpatterns += static(MEDIA_URL, document_root=MEDIA_ROOT)
