backend/requirements.txt
requests==2.26.0
Django==3.2.25
djangorestframework==3.12.4
django-filter==23.5
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
Pillow==10.3.0
djoser==2.2.2
drf-extra-fields==3.7.0